Fixes for python sdk
