+++ sdk-py/examples/question_answer_demo.py
@@ -1,10 +1,12 @@
-from indexify import Index, TextChunk
+from indexify import Index, Repository, TextChunk, DEFAULT_INDEXIFY_URL
 from datasets import load_dataset
 
 
 class DemoQA:
+
     def __init__(self):
-        self.idx = Index(Index.DEFAULT_INDEXIFY_URL, "default/default")
+        self.repository = Repository(DEFAULT_INDEXIFY_URL, "default/default")
+        self.idx = Index(DEFAULT_INDEXIFY_URL, "default/default")
 
     def execute(self):
         # Add All Wikipedia articles
@@ -14,7 +16,7 @@ class DemoQA:
             context: str = datasets[i]["context"]
             question = datasets[i]["question"]
             answers = datasets[i]["answers"]
-            self.idx.add(TextChunk(context))
+            self.repository.add(TextChunk(context))
             q_a_all.append((question, answers))
         for q_a in q_a_all:
             question = q_a[0]
