+++ sdk-py/indexify/__init__.py
@@ -1,4 +1,7 @@
-from index import Index, Message, TextChunk
-from memory import Memory
+from .index import Index, Message, TextChunk
+from .memory import Memory
+from .repository import Repository
 
-__all__ = ["Index", "Memory", "Message", "TextChunk"]
+DEFAULT_INDEXIFY_URL = "http://localhost:8900"
+
+__all__ = ["Index", "Memory", "Message", "TextChunk", "Repository", "DEFAULT_INDEXIFY_URL"]
