+++ sdk-py/indexify/data_containers.py
@@ -1,5 +1,6 @@
 from enum import Enum
 from typing import Optional, List
+from dataclasses import dataclass, field
 
 
 class TextSplitter(str, Enum):
@@ -10,26 +11,24 @@ class TextSplitter(str, Enum):
     def __str__(self) -> str:
         return self.value.lower()
 
-
+@dataclass
 class TextChunk:
-    def __init__(self, text: str, metadata: Optional[dict] = {}):
-        self.text = text
-        self.metadata = metadata
+    text: str
+    metadata: dict[str, any] = field(default_factory=dict)
 
     def to_dict(self):
         return {"text": self.text, "metadata": self.metadata}
 
-
+@dataclass
 class Message:
-    def __init__(self, role: str, text: str, metadata: dict = {}):
-        self.role = role
-        self.text = text
-        self.metadata = metadata
+    role: str
+    text: str
+    metadata: dict[str, any] = field(default_factory=dict)
 
     def to_dict(self):
         return {"role": self.role, "text": self.text, "metadata": self.metadata}
 
-
+@dataclass
 class SearchChunk:
     index: str
     query: str
@@ -38,6 +37,6 @@ class SearchChunk:
     def to_dict(self):
         return {"index": self.index, "query": self.query, "k": self.k}
 
-
+@dataclass
 class SearchResult:
     results: List[TextChunk]
